# Frontend/maintenance-frontend/src/app/service/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/api/Utilisateur/update":{"put":{"tags":["utilisateur-controller"],"operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Contrat/update/{id}":{"put":{"tags":["contrat-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contrat"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Contrat"}}}}}}},"/login":{"post":{"tags":["login-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/Utilisateur/reset-password":{"post":{"tags":["utilisateur-controller"],"operationId":"resetPassword","parameters":[{"name":"newPassword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/Utilisateur/add":{"post":{"tags":["utilisateur-controller"],"operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Utilisateur/addUserWithTempPassword":{"post":{"tags":["utilisateur-controller"],"operationId":"addUserWithTempPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Utilisateur/add-admin":{"post":{"tags":["utilisateur-controller"],"operationId":"addAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Imprimante/import-excel":{"post":{"tags":["imprimante-controller"],"operationId":"importImprimantesFromExcel","parameters":[{"name":"contratId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/Imprimante/assign-multiple":{"post":{"tags":["imprimante-controller"],"operationId":"assignImprimantesToContrat","parameters":[{"name":"imprimanteIds","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}},{"name":"contratId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Imprimante"}}}}}}}},"/api/Imprimante/addImprimante":{"post":{"tags":["imprimante-controller"],"operationId":"addImprimante","parameters":[{"name":"contratId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Imprimante"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Imprimante"}}}}}}},"/api/Contrat/save":{"post":{"tags":["contrat-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contrat"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Contrat"}}}}}}},"/api/Contrat/renouveler/{id}":{"post":{"tags":["contrat-controller"],"operationId":"renouvelerContrat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contrat"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Contrat"}}}}}}},"/api/Contrat/delete/{id}":{"post":{"tags":["contrat-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/Utilisateur/getbyid/{id}":{"get":{"tags":["utilisateur-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Utilisateur/getbyemail/{email}":{"get":{"tags":["utilisateur-controller"],"operationId":"getUserByEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Utilisateur"}}}}}}},"/api/Utilisateur/getallUser":{"get":{"tags":["utilisateur-controller"],"operationId":"retrieveAllUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Utilisateur"}}}}}}}},"/api/Imprimante/getImprimante":{"get":{"tags":["imprimante-controller"],"operationId":"getImprimante","parameters":[{"name":"imprimanteId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Imprimante"}}}}}}},"/api/Imprimante/getAllImprimante":{"get":{"tags":["imprimante-controller"],"operationId":"getAllImprimante","parameters":[{"name":"contratId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Imprimante"}}}}}}}},"/api/Contrat/getContratsHistorie":{"get":{"tags":["contrat-controller"],"operationId":"getContratsHistorie","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Contrat"}}}}}}}},"/api/Contrat/findById/{id}":{"get":{"tags":["contrat-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Contrat"}}}}}}},"/api/Contrat/findAll":{"get":{"tags":["contrat-controller"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Contrat"}}}}}}}},"/api/Utilisateur/delete/{idUsers}":{"delete":{"tags":["utilisateur-controller"],"operationId":"removeUser","parameters":[{"name":"idUsers","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/Imprimante/deleteImprimante":{"delete":{"tags":["imprimante-controller"],"operationId":"deleteImprimante","parameters":[{"name":"imprimanteId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/Imprimante/delete-by-contrat/{contratId}":{"delete":{"tags":["imprimante-controller"],"operationId":"deleteAllImprimantesByContratId","parameters":[{"name":"contratId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Contrat":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numeroContrat":{"type":"string"},"dateDebut":{"type":"string","format":"date"},"dateFin":{"type":"string","format":"date"},"statutContrat":{"type":"string","enum":["ACTIF","EXPIRE","RENOUVELE"]},"conditions_contrat":{"type":"string"},"client":{"$ref":"#/components/schemas/Utilisateur"},"facture":{"$ref":"#/components/schemas/Facture"},"intervention":{"$ref":"#/components/schemas/Intervention"},"contratPrecedent":{"$ref":"#/components/schemas/Contrat"}}},"Facture":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"montant":{"type":"number","format":"double"},"statutFacture":{"type":"string","enum":["PAYEE","NON_PAYEE"]},"contrat":{"$ref":"#/components/schemas/Contrat"}}},"Imprimante":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"marque":{"type":"string"},"modele":{"type":"string"},"emplacement":{"type":"string"},"numeroSerie":{"type":"string"},"contrat":{"$ref":"#/components/schemas/Contrat"},"interventions":{"type":"array","items":{"$ref":"#/components/schemas/Intervention"}}}},"Intervention":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"typeProbleme":{"type":"string"},"description":{"type":"string"},"dateDemande":{"type":"string","format":"date"},"dateIntervention":{"type":"string","format":"date"},"statutIntervention":{"type":"string","enum":["EN_ATTENTE","EN_COURS","TERMINEE"]},"priorite":{"type":"string"},"cout":{"type":"number","format":"double"},"imprimante":{"$ref":"#/components/schemas/Imprimante"},"technicien":{"$ref":"#/components/schemas/Utilisateur"},"contrat":{"$ref":"#/components/schemas/Contrat"}}},"Notification":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"message":{"type":"string"},"date":{"type":"string","format":"date"},"utilisateur":{"$ref":"#/components/schemas/Utilisateur"}}},"Utilisateur":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"prenom":{"type":"string"},"imageUrl":{"type":"string"},"email":{"type":"string"},"telephone":{"type":"string"},"motDePasse":{"type":"string"},"role":{"type":"string","enum":["ADMIN","TECHNICIAN","CLIENT"]},"mustResetPassword":{"type":"boolean"},"interventions":{"type":"array","items":{"$ref":"#/components/schemas/Intervention"}},"notifications":{"type":"array","items":{"$ref":"#/components/schemas/Notification"}}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"motDePasse":{"type":"string"}}}}}}